Feat: add uvx support by renaming src to zabbix_mcp package
