+++ pyproject.toml
@@ -36,7 +36,11 @@ Issues = "https://github.com/mpeirone/zabbix-mcp-server/issues"
 Documentation = "https://github.com/mpeirone/zabbix-mcp-server#readme"
 
 [project.scripts]
-zabbix-mcp = "src.zabbix_mcp_server:main"
+zabbix-mcp = "zabbix_mcp.zabbix_mcp_server:main"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["zabbix_mcp*"]
 
 [build-system]
 requires = ["setuptools>=61"]
